+++ examples/hello_world/android/app/src/main/kotlin/org/leanflutter/examples/hello_world/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.hello_world
+package org.leanflutter.examples.hello_world
 
 import io.flutter.embedding.android.FlutterActivity