+++ packages/flutter_app_builder/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.7
+
+- Change parameter `cleanOnceBeforeBuild` to `cleanBeforeBuild`
+
 ## 0.1.2
 
 - `dart-define` arg support multi values.